+++ src/v2/providers/alerts/alerts.ts
@@ -94,7 +94,7 @@ export interface FirebaseAlertOptions extends options.EventHandlerOptions {
   /**
    * Region where functions should be deployed.
    */
-  region?: options.SupportedRegion | string;
+  region?: options.SupportedRegion | string | Expression<string> | ResetValue;
 
   /**
    * Amount of memory to allocate to a function.